agent: backhaul_manager: refactor tlvSupportedFreqBand handling
The handling of the 2.4GHz and 5GHz bands were exactly the same,
except for the log messages and freq_type.

Since there are already 3 different frequency bands to handle, use a
switch-case to handle each of them.

If the band is supported, assign a band name and a freq_type. If it's
not supported, log it and return.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>