agent: backhaul_manager: call finalize if state is already operational
Previously, when an auto-config reply was received for a specific
band, finalize_slaves_connect_state() would only be called if the
backhaul was already operational on slave registration. This is
because if it wasn't the state change to OPERATIONAL would already
call finalize_slaves_connect_state().

There is a case where finalize_slaves_connect_state() wouldn't be
called for all the slaves though: if the auto-config message for one
of the slave arrived after the FSM went to OPERATIONAL state.

In that case, the FSM state change would cause the interface for which
the auto-config reply wasn't received to be skipped with the following
message: "controller not discovered, skipping...".
The finalize_slaves_connect_state() function would then never be
called, and the slave wouldn't be operational.

Add a condition to also call finalize_slaves_connect_state() if the
current state is operational, but wasn't at the time of the slave
registration.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>