easylogging: resolveFilename: use same file for all debug levels
For each debug level the log file name is calculated separately.
Normally, it's always the same name, so the log file will be opened
only once. However, if the second happens to tick by in the middle of
the initialisation of the logging, we will have two different
log files with 1 second difference, and half the log ends up
in one log file, the other half in the other log file.

In the case of that CI job, it lead to a failure because
we grep the log file to check the status of the agent,
and we were grepping the wrong log file.

This commit ensures that the time stamp is cached and shared
between all the debug levels.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@gmail.com>