bwl: Fix not unique BSSID mac address
Currently, a radio mac on the dummy is equal to the first BSSID mac on that radio.
This is wrong since the mac address must be unique values on the controller database
because they are saved on unrodered_map.
Fix the first BSSID mac assignment on dummy BWL to not be equal to the radio mac.
Also, fix it on "sta_bss_steer" function that assigned the radio mac not
the bssid mac.

Fix the test scripts too, to use the bssid mac instead of radio mac.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>