agent: Modify ap_manager constructor
Add "iface" string and the "logging" class, an arguments to ap_manager constructor.
While at it:
- Change the slave_uds argument to const reference, and move its initialization to
the constructor body.
- Change "bss_steer_valid_int" and "bss_steer_imminent_valid_int" members to be
initialized on the class definition instead of the constructor.
Also, move the default values from beerocks_defines to the ap_manager class since
they are not being used from somewhere else.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>