agent: Move sending of ACTION_MONITOR_JOINED_NOTIFICATION
ACTION_MONITOR_JOINED_NOTIFICATION is currently being sent on the init function of the monitor.
This is not making much sense since at the moment the monitor is not attached and cannot do
anything.
Move the sending of the ACTION_MONITOR_JOINED_NOTIFICATION message to right after the monitor
has finished the attachment process.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>