tests: added two unit test applications
This commit adds two test target that checks
the conversions between vs and 1905
for beacon metrics query and response.
it is used during development and can also
be used during regualr builds as a test.
currently the DUT, the conversion cpp function
implementation is compiled twice:
* as part of the son-slave
* as part of the test application
The best would be to create a library for conversion.

Signed-off-by: Ran regev <ran.regev@devalore.com>