tests: environment.py: support multiple log files
As part of the unified agent refactoring, the per-radio logs are now
spread over three files instead of two: agent_wlanX, ap_manager_wlanX
and monitor_wlanX. The agent_wlanX log file will eventually be removed
again, and we should look for that information in ap_manager_wlanX
instead.

For the time being, however, it's easier to simply look in both of them.
When unified agent is done, the cases where we look in agent_wlanX can
instead be converted into check_log on the agent instead of the radio.

Until that time, as a workaround, support two log files on the radio:
agent and ap_manager. Update _docker_wait_for_log to support multiple
programs instead of a single one, and update the callers to pass a list
instead of a single entry.

This change should be reverted again when unified agent is done.
Therefore, add a big comment before _docker_wait_for_log that says so.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>