controller: DCS task: revert changes
The old "Scan on all channels" worked from the DCS task.
This was not the best course of action as it failed to take into
consideration the "supported channel" validation when setting the pool
and the "result in pool" validation when the results are stored in the
DB.

Revert the changes made in 47d56372.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>