controller: Add "scan on all channels" support
As part of the new "Scan on all channels" the client can requests a scan
with a specific pool that will indicate they wish to run a scan on all
the radio's supported channels.
When the client sends a channel pool with a single value of '0' it will
indicate that the client wants to use the "Scan on all channels"
feature.

When setting the channels pool in son_management, supported scanning on
all channels in the pool matches the criteria.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>