.gitlab-ci.yml: add ubuntu xenial to the registry
Commit 1635545c11b2b4ada215b51341aedab590967e41 added support for
building ubuntu xenial (16.04) as part of CI, but mistakingly used
docker.io registry.
Fix this now, by adding it to the images build jobs in .gitlab-ci.yml and
adding dependency on it for the build-in-docker-ubuntu-xenial so it will
be automatically built and pushed to the the gitlab registry as part of
CI.

While we're at it, move the Dockerfile to ubuntu/xenial and add a
Dockerfile for ubuntu bionic (18.04). In a later commit we will use this
one instead of the top level ./tools/docker/builder/Dockerfile - it was
originally designed to support multiple base images, but  it is not
realistic to support all variants in a single Dockerfile since packages
may differ just like with completely different distributions.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>