tests: sniffer.py: use dumpcap instead of tcpdump
It seems that tcpdump sometimes doesn't process packets while they have
been captured by the kernel. This behaviour wasn't observed with
dumpcap. In addition, dumpcap is lighter-weight than tcpdump (since it
doesn't do analysis, just capture).

tshark was considered as well, but is unfortunately racy because it
doesn't print anything when the capture has been started - it only
prints something just before starting the capture.

dumpcap doesn't need the equivalent of the -U option, since it never
buffers the output file. On the other hand, it is useful to add the -q
option to suppress the continuous Packets output. This actually removes
the need to close stderr since nothing will be printed until the process
is killed, but that's something that can be changed later if required.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>