ci/owncloud: add a default webdav options file
although the script will utilize .netrc for getting the credentials
required for logging into owncloud, for whichever reason one might want,
it will also support getting the credentials from an external file.
this file serves as a skeleton. The format is identical to the python
dictionary utilized by the webdav module.

Signed-off-by: Gal Wiener <gal.wiener@intel.com>