btlvf: extend cACTION_BACKHAUL_CLIENT_ASSOCIATED_NOTIFICATION
Preparative commit.
Currently on client connected notification from the AP manager, two
messages are sent - one VS message to the backhaul manager which needs
this information for topology messages, and one to the controller since
the agent as a whole should send the notification whenever a client
joins. Since we anyway need to send the message to the backhaul manager,
we better use the VS message with all the information and generate the
topology notification as we do with other topology message in the
backhaul manager.

Add beerocks::message::sRadioCapabilities and vap_id to
cACTION_BACKHAUL_CLIENT_ASSOCIATED_NOTIFICATION.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>