btlvf: extend cACTION_BACKHAUL_CLIENT_DISCONNECTED_NOTIFICATION
Preparative commit.
Extend cACTION_BACKHAUL_CLIENT_DISCONNECTED_NOTIFICATION with vap_id,
disconnect reason, source and type which are needed by the prplmesh controller.
This is added to the backhaul notification since it will in turn build
the TOPOLOGY_NOTIFICATION CMDU which will hold the VS TLV with these
variables, instead of the son slave which does it now and sends two
messages - one to the backhaul and one to the controller.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>